Add support to load binary patch
The binary patch is converted to a relocatable object and linked with
the driver and is used to load the patch to RPU. We append "_shipped" to
avoid kernel building the object.

In Zephyr we use "incbin" approach with inline assembly which doesn't
seem to work for RPI, hence we use "ld".

Implements SHEL-2110.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>